requirements.txt
aioredis==1.3.1
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.1
cssselect2==0.6.0
Cython==0.29.28
daphne==3.0.2
dj-database-url==1.0.0
Django==4.0.3
django-channels==0.7.0
django-channels-handlers==0.2.2
django-filter==21.1
django-heroku==0.3.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
fonttools==4.33.3
future==0.18.2
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
lxml==4.8.0
msgpack==1.0.3
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.10
oscrypto==1.3.0
Pillow==9.1.0
pisa==3.0.33
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydyf==0.1.2
pyHanko==0.13.0
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.0.0
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
scipy==1.8.0
service-identity==21.1.0
six==1.16.0
smart-open==6.0.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.64.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
weasyprint==54.3
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.7
xlwt==1.3.0
zope.interface==5.4.0
zopfli==0.2.1
